made some adjustments to the signup View controller. I am hoping I dont get errors or merge conflicts.
